requirements.txt
spotipy==2.23.0
rich==13.7.0
python-dotenv==1.0.1 